Revert "merge with main"
This reverts commit 77a491a86ed98c28c0ecd33d6dded24f01c94be7.

+++ Makefile
@@ -23,10 +23,11 @@ ROM_REVISION ?= 1
 
 SUBPROJ_DIR := subprojects
 
-MESON_VER := 1.12.0
+MESON_VER := 1.10.0
 MESON_DIR := $(SUBPROJ_DIR)/meson-$(MESON_VER)
+MESON_SUB := $(MESON_DIR)/meson.py
 
-MESON ?= $(MESON_DIR)/meson.py
+MESON ?= $(MESON_SUB)
 NINJA ?= ninja
 GIT ?= git
 
@@ -118,8 +119,12 @@ distclean:
 
 purge: distclean
 	rm -rf $(SKREW_DIR)
+ifeq ($(MESON),$(MESON_SUB))
 	! test -f $(MESON) || $(MESON) subprojects purge --confirm
 	rm -rf $(MESON_DIR)
+else
+	$(MESON) subprojects purge --confirm
+endif
 
 update: meson skrewup
 	$(MESON) subprojects update || true
@@ -143,9 +148,12 @@ $(BUILD)/build.ninja: | $(BUILD) $(SKREW_EXE) meson
 $(BUILD):
 	mkdir -p -- $(BUILD)
 
-meson: $(MESON)
+meson: ;
+ifeq ($(MESON),$(MESON_SUB))
+meson: $(MESON_SUB)
+endif
 
-$(MESON):
+$(MESON_SUB):
 	$(GIT) clone --depth=1 -b $(MESON_VER) https://github.com/mesonbuild/meson $(@D)
 
 skrew: $(SKREW_EXE)
